Feat : add oauth2 login, add modification member profile

+++ src/main/java/com/bit/healthpartnerboot/entity/MemberHistory.java
@@ -23,9 +23,9 @@ public class MemberHistory {
     @Column(name = "history_seq")
     private Integer seq;
 
-    private Integer height;
+    private Float height;
 
-    private Integer weight;
+    private Float weight;
 
     private Float bmi;
